requirements.txt
dhanhq
django-admin
django
django-allauth
python-dotenv
channels==3.0.5
websocket